Y.-Y Tai check in for the fixed link and cleaning for example folder.
The link of the TBM3 library is now fixed to a single include folder:
  [system path]/TBM3/include

Users should include “tbm3.h” to start with.

At this commit, the code is workable and one can try the example in :
TBM3/example/main_cubic.cpp
to test it.

The Wannier part will be added in the next few commit.